+++ webapp/py/fonteditor.py
@@ -174,27 +174,3 @@ def CurrentFont():
 #         print(f"Error saving font: {e}")
 #         return False
 
-
-def GetOpentypeFeatureInfo():
-    """
-    Get information about OpenType features, including which are discretionary
-    and which should be on by default.
-
-    Returns:
-        dict: Dictionary with feature information including:
-            - 'default_on': List of features that should be on by default
-            - 'default_off': List of features that should be off by default
-            - 'descriptions': Dictionary mapping feature tags to descriptions
-
-    Example:
-        >>> info = GetOpentypeFeatureInfo()
-        >>> print(info['default_on'])
-        ['calt', 'clig', 'liga', 'kern', 'cpsp', 'locl']
-    """
-    import opentype
-
-    return {
-        "default_on": list(opentype.DEFAULT_ON_FEATURES),
-        "default_off": list(opentype.DEFAULT_OFF_FEATURES),
-        "descriptions": opentype.FEATURE_DESCRIPTIONS,
-    }